+++ src/vs/workbench/api/test/node/extHostHooks.test.ts
@@ -9,8 +9,8 @@ import { ensureNoDisposablesAreLeakedInTestSuite } from '../../../../base/test/c
 import { NullLogService } from '../../../../platform/log/common/log.js';
 import { NodeExtHostHooks } from '../../node/extHostHooksNode.js';
 import { IHookCommandDto, MainThreadHooksShape } from '../../common/extHost.protocol.js';
-import { HookCommandResultKind } from '../../../contrib/chat/common/hooksCommandTypes.js';
-import { IHookResult } from '../../../contrib/chat/common/hooksTypes.js';
+import { HookCommandResultKind } from '../../../contrib/chat/common/hooks/hooksCommandTypes.js';
+import { IHookResult } from '../../../contrib/chat/common/hooks/hooksTypes.js';
 import { IExtHostRpcService } from '../../common/extHostRpcService.js';
 import { CancellationToken } from '../../../../base/common/cancellation.js';
