+++ src/vs/workbench/contrib/chat/common/hooks/hooksExecutionService.ts
@@ -3,16 +3,16 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { createDecorator } from '../../../../platform/instantiation/common/instantiation.js';
-import { URI } from '../../../../base/common/uri.js';
-import { HookType, HookTypeValue, IChatRequestHooks, IHookCommand } from './promptSyntax/hookSchema.js';
-import { IDisposable, toDisposable } from '../../../../base/common/lifecycle.js';
-import { ILogService } from '../../../../platform/log/common/log.js';
-import { CancellationToken } from '../../../../base/common/cancellation.js';
-import { StopWatch } from '../../../../base/common/stopwatch.js';
-import { Extensions, IOutputChannelRegistry, IOutputService } from '../../../services/output/common/output.js';
-import { Registry } from '../../../../platform/registry/common/platform.js';
-import { localize } from '../../../../nls.js';
+import { createDecorator } from '../../../../../platform/instantiation/common/instantiation.js';
+import { URI } from '../../../../../base/common/uri.js';
+import { HookType, HookTypeValue, IChatRequestHooks, IHookCommand } from '../promptSyntax/hookSchema.js';
+import { IDisposable, toDisposable } from '../../../../../base/common/lifecycle.js';
+import { ILogService } from '../../../../../platform/log/common/log.js';
+import { CancellationToken } from '../../../../../base/common/cancellation.js';
+import { StopWatch } from '../../../../../base/common/stopwatch.js';
+import { Extensions, IOutputChannelRegistry, IOutputService } from '../../../../services/output/common/output.js';
+import { Registry } from '../../../../../platform/registry/common/platform.js';
+import { localize } from '../../../../../nls.js';
 import {
 	HookCommandResultKind,
 	IHookCommandInput,
